[fix] Change project structure
