+++ .gitignore
@@ -28,6 +28,7 @@ yarn-error.log
 .history/*
 
 # Miscellaneous
+.angular
 /.angular/cache
 .sass-cache/
 /connect.lock
