+++ apps/en/src/app/base/app.component.ts
