+++ apps/en/src/app/base/error/error.component.scss
