+++ apps/en/src/app/base/posts/posts.component.scss
