+++ apps/en/src/app/core/common/const/route-paths.ts
