+++ apps/en/src/app/features/posts/ui/common/components/post-card/post-card.component.scss
