+++ apps/en/src/app/ui/dumb/header/header.component.scss
@@ -1,0 +1,31 @@
+@use '../../../../../../../libs/assets/src/lib/styles/mixins';
+
+.header {
+    background-color: #242424;
+    color: #855ec8;
+    padding: 15px 20px;
+}
+
+.nav-list {
+    @include mixins.flex(row, center, center);
+    list-style: none;
+    gap: 20px;
+}
+
+.nav-link {
+    color: #855ec8;
+    text-decoration: none;
+    font-size: 1rem;
+    padding: 5px 10px;
+    transition: 0.3s ease;
+
+    &:hover {
+        transform: scale(1.02);
+        text-decoration: underline;
+        color: lighten(#855ec8, 5%);
+    }
+}
+
+.active-link {
+    color: lighten(#855ec8, 30%);
+}
