+++ apps/en/src/main.ts
@@ -1,0 +1,6 @@
+import {bootstrapApplication} from '@angular/platform-browser';
+import {appConfig} from './app/infrastructure/configs/app.config';
+import {AppComponent} from "./app/base/app.component";
+
+bootstrapApplication(AppComponent, appConfig)
+    .catch((err) => console.error(err));
