+++ libs/assets/src/lib/styles/styles.scss
