--- src/apps/en/src/app/base/app.component.scss
@@ -1,107 +1,0 @@
-@use '../core/common/styles/mixins';
-
-.error-wrapper {
-    height: 100vh;
-    @include mixins.flex(row, center, center)
-}
-
-.loader-wrapper {
-    height: 100vh;
-    @include mixins.flex(row, center, center)
-}
-
-// нашел готовый лоадер
-.loader {
-    font-size: 10px;
-    width: 1em;
-    height: 1em;
-    border-radius: 50%;
-    position: relative;
-    text-indent: -9999em;
-    animation: mulShdSpin 1.1s infinite ease;
-    transform: translateZ(0);
-}
-
-@keyframes mulShdSpin {
-    0%,
-    100% {
-        box-shadow: 0em -2.6em 0em 0em #9c27b0,
-        1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2),
-        2.5em 0em 0 0em rgba(156, 39, 176, 0.2),
-        1.75em 1.75em 0 0em rgba(156, 39, 176, 0.2),
-        0em 2.5em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em 1.8em 0 0em rgba(156, 39, 176, 0.2),
-        -2.6em 0em 0 0em rgba(156, 39, 176, 0.5),
-        -1.8em -1.8em 0 0em rgba(156, 39, 176, 0.7);
-    }
-    12.5% {
-        box-shadow: 0em -2.6em 0em 0em rgba(156, 39, 176, 0.7),
-        1.8em -1.8em 0 0em #9c27b0,
-        2.5em 0em 0 0em rgba(156, 39, 176, 0.2),
-        1.75em 1.75em 0 0em rgba(156, 39, 176, 0.2),
-        0em 2.5em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em 1.8em 0 0em rgba(156, 39, 176, 0.2),
-        -2.6em 0em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em -1.8em 0 0em rgba(156, 39, 176, 0.5);
-    }
-    25% {
-        box-shadow: 0em -2.6em 0em 0em rgba(156, 39, 176, 0.5),
-        1.8em -1.8em 0 0em rgba(156, 39, 176, 0.7),
-        2.5em 0em 0 0em #9c27b0,
-        1.75em 1.75em 0 0em rgba(156, 39, 176, 0.2),
-        0em 2.5em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em 1.8em 0 0em rgba(156, 39, 176, 0.2),
-        -2.6em 0em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2);
-    }
-    37.5% {
-        box-shadow: 0em -2.6em 0em 0em rgba(156, 39, 176, 0.2),
-        1.8em -1.8em 0 0em rgba(156, 39, 176, 0.5),
-        2.5em 0em 0 0em rgba(156, 39, 176, 0.7),
-        1.75em 1.75em 0 0em #9c27b0,
-        0em 2.5em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em 1.8em 0 0em rgba(156, 39, 176, 0.2),
-        -2.6em 0em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2);
-    }
-    50% {
-        box-shadow: 0em -2.6em 0em 0em rgba(156, 39, 176, 0.2),
-        1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2),
-        2.5em 0em 0 0em rgba(156, 39, 176, 0.5),
-        1.75em 1.75em 0 0em rgba(156, 39, 176, 0.7),
-        0em 2.5em 0 0em #9c27b0,
-        -1.8em 1.8em 0 0em rgba(156, 39, 176, 0.2),
-        -2.6em 0em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2);
-    }
-    62.5% {
-        box-shadow: 0em -2.6em 0em 0em rgba(156, 39, 176, 0.2),
-        1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2),
-        2.5em 0em 0 0em rgba(156, 39, 176, 0.2),
-        1.75em 1.75em 0 0em rgba(156, 39, 176, 0.5),
-        0em 2.5em 0 0em rgba(156, 39, 176, 0.7),
-        -1.8em 1.8em 0 0em #9c27b0,
-        -2.6em 0em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2);
-    }
-    75% {
-        box-shadow: 0em -2.6em 0em 0em rgba(156, 39, 176, 0.2),
-        1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2),
-        2.5em 0em 0 0em rgba(156, 39, 176, 0.2),
-        1.75em 1.75em 0 0em rgba(156, 39, 176, 0.2),
-        0em 2.5em 0 0em rgba(156, 39, 176, 0.5),
-        -1.8em 1.8em 0 0em rgba(156, 39, 176, 0.7),
-        -2.6em 0em 0 0em #9c27b0,
-        -1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2);
-    }
-    87.5% {
-        box-shadow: 0em -2.6em 0em 0em rgba(156, 39, 176, 0.2),
-        1.8em -1.8em 0 0em rgba(156, 39, 176, 0.2),
-        2.5em 0em 0 0em rgba(156, 39, 176, 0.2),
-        1.75em 1.75em 0 0em rgba(156, 39, 176, 0.2),
-        0em 2.5em 0 0em rgba(156, 39, 176, 0.2),
-        -1.8em 1.8em 0 0em rgba(156, 39, 176, 0.5),
-        -2.6em 0em 0 0em rgba(156, 39, 176, 0.7),
-        -1.8em -1.8em 0 0em #9c27b0;
-    }
-}
