--- src/apps/en/src/app/base/error/error.component.ts
@@ -1,24 +1,0 @@
-import {Component} from '@angular/core';
-import {Location} from '@angular/common';
-
-import { RouterLink } from "@angular/router"
-import {ROUTE_PATHS} from "../../core/common/constants/route-paths";
-
-@Component({
-    selector: 'apps-error',
-    imports: [
-        RouterLink
-    ],
-    templateUrl: './error.component.html',
-    styleUrl: './error.component.scss'
-})
-export class ErrorComponent {
-    ROUTE = ROUTE_PATHS;
-
-    constructor(private location: Location) {}
-
-    goBack(): void {
-        // Почему-то при клике обновляет стр., но возвращает. Делал как в доке
-        this.location.back()
-    }
-}
