+++ tsconfig.spec.json
@@ -9,7 +9,7 @@
     ]
   },
   "include": [
-    "src/**/*.spec.ts",
-    "src/**/*.d.ts"
+    "apps/**/*.spec.ts",
+    "apps/**/*.d.ts"
   ]
 }
